requirements.txt
feedparser
html2text
langdetect
nltk
sklearn
facebook-sdk
pickledb
numpy
scipy